Make all functions return by reference (no stdout), and use dynamic variables to avoid variable collisions
All functions now return values through a provided reference variable name instead of
stdout.  This way, no subshells are ever used so functions can assign values to
variables that exist in the scope of the caller.

Dynamic variables are used in all intermediate functions because the assumption that a
specific variable name will be unused in the scope of the caller cannot be made, and causes
overwriting if it is.